Avoid git reset for the parent directory

+++ plugin/jetpack.vim
@@ -230,6 +230,10 @@ function! jetpack#clean_plugins() abort
     return
   endif
   for [pkg_name, pkg] in items(s:declared_packages)
+    if !isdirectory(pkg.path . '/.git')
+      call delete(pkg.path, 'rf')
+      continue
+    endif
     if isdirectory(pkg.path)
       call system(printf('git -C %s reset --hard', pkg.path))
       let branch = trim(system(printf('git -C %s rev-parse --abbrev-ref %s', pkg.path, pkg.commit)))
